pyproject.toml
[tool.poetry]
name = "SWE_HW1"
version = "0.1.0"
description = ""
authors = ["Jack Krub"]
packages = [
    { include = "command_template" },
]

[tool.poetry.dependencies]
python = ">=3.7,<4.0"
flask = "^2.2.2"
mysql = "^0.0.3"
flask-mysqldb = "^1.0.1"
flask-bootstrap = "^3.3.7.1"
pymysql = "^1.0.2"
flask-wtf = "^1.1.1"
requests = "^2.28.2"
chardet = "^5.1.0"
cryptography = "^39.0.0"


[tool.poetry.dev-dependencies]
pre-commit = "^2.20.0"
flake8 = "^5.0.1"
pytest = "^7.1.2"
pytest-cov = "^3.0.0"
ipykernel = "^6.15.1"
flake8-annotations = "^2.9.1"
flake8-eradicate = "^1.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
